+++ spec/rubocop/cop/rspec/describe_class_spec.rb
@@ -44,6 +44,11 @@ describe RuboCop::Cop::RSpec::DescribeClass do
     expect(cop.offenses).to be_empty
   end
 
+  it 'ignores routing specs' do
+    inspect_source(cop, "describe 'my new route', type: :routing do; end")
+    expect(cop.offenses).to be_empty
+  end
+
   it "doesn't blow up on single-line describes" do
     inspect_source(cop, 'describe Some::Class')
     expect(cop.offenses).to be_empty